Add preserve content filenames option for CSV
This provides the option to retain the source filename for the
metadata and copied content. It also means that if the input
and output directories match (and filenames are preserved) then
it will act in metadata only mode i.e. no copying of content
will take place which will speed things up for large batches.