Implements shutdown hooks.
https://github.com/MarcusBarnes/mik/issues/156

Documentation snippet:

Similar to post-write hooks, but as the name implies, shutdown hooks run just before mik exits.
Shutdown hooks also differ in that they are run sequentially. That is, the first shutdown hook must
exit before the second is invoked.

Register shutdown hooks in .ini in the `[WRITER]` section:

	 shutdownhooks[] = "/usr/bin/php extras/scripts/shutdownhooks/shutitdown.php"