Adds a real-world shutdown hook for testing.
This script is an adaptation of a post-write hook of the same
name. Because the post-write version opened a new saxon process
for each file, it was too easy for post-write to eat all of a
system's resources.

This incarnation leverages saxon's ability to operate on entire
directories. Testing note: saxon will choke a little if the input
directory contains files other than valid .xml.